wip
now the repl menu is keeping state around, but not watching that the editor changed (?)

There's a disconnect between selected kernel and the kernel that's running in the editor.
